fix: post stock
