+++ src/models/Stock.ts
@@ -1,3 +1,4 @@
+import * as createError from 'http-errors';
 import {
   AllowNull,
   AutoIncrement,
@@ -11,7 +12,7 @@ import {
   Unique
 } from 'sequelize-typescript';
 import { ERROR_CODES, HTTP_ERRORS } from '../constants';
-import { validateDuplicateFields } from './index';
+const { BadRequest } = createError;
 
 const { STOCK_DUPLICATE_CODE } = ERROR_CODES;
 
@@ -19,6 +20,40 @@ const duplicateFields = {
   stock_code: {
     field: 'stock_code',
     error: HTTP_ERRORS[STOCK_DUPLICATE_CODE].MESSAGE
+  }
+};
+
+export interface IStockFieldValidateDup {
+  instance: Stocks;
+  field: 'stock_code';
+  error: string;
+}
+const validateDuplicate = async (
+  options: IStockFieldValidateDup
+): Promise<void> => {
+  const { instance, field, error } = options;
+  if (instance.changed(field) && instance.previous(field) !== instance[field]) {
+    const query = { where: { [field]: instance[field] } };
+    const stock = await Stocks.findOne(query);
+
+    if (stock) throw new BadRequest(error);
+  }
+
+  return;
+};
+
+const validateDuplicateFields = async (instance: Stocks): Promise<void> => {
+  try {
+    const keys = Object.keys(duplicateFields);
+
+    for (const key of keys) {
+      const { field, error } = duplicateFields[key];
+      await validateDuplicate({ instance, field, error });
+    }
+
+    return;
+  } catch (error) {
+    throw error;
   }
 };
 
@@ -65,7 +100,7 @@ export class Stocks extends Model<Stocks> {
   @BeforeCreate
   @BeforeUpdate
   static async beforeSaveInstance(instance: Stocks) {
-    await validateDuplicateFields(instance, Stocks, duplicateFields);
+    await validateDuplicateFields(instance);
   }
 
   @BeforeUpdate
